Correctly bring up F2FS support
In order for a correct Filesystem bringup, there were missing a few flags and options.
That is, was causing a correct encryption while making the device unlockable. Those missing flags'll fix the problem!
Thanks to ElPaablo and Oleks especially (It's extracted from his fix in the CrDroid for Mi9SE's group)